dom/canvas/test/webgl-mochitest/ensure-exts/test_common.html

<!DOCTYPE HTML>
<html>
  <head>
    <meta charset='utf-8'/>
    <script src='/tests/SimpleTest/SimpleTest.js'></script>
    <link rel='stylesheet' href='/tests/SimpleTest/test.css'>
  </head>
  <body>
    <script>

'use strict';

var c = document.createElement('canvas');
var gl = c.getContext('experimental-webgl');

function ensureExt(name) {
  ok(gl.getExtension(name), 'Should have extension ' + name + '.');
}

function ensureNoExt(name) {
  ok(!gl.getExtension(name), 'Should not have extension ' + name + '.');
}

do {
  if (!gl)
    break;

  // These aren't all guaranteed, but they're common to all our test slaves.
  // If you're adding a slave config that is missing one of these, comment the line out
  // and split it into its own test.

  // Implemented. (commented out if not test-slave-universal)
  //ensureExt('ANGLE_instanced_arrays');
  //ensureExt('EXT_blend_minmax');
  //ensureExt('EXT_color_buffer_half_float');
  //ensureExt('EXT_frag_depth');
  //ensureExt('EXT_shader_texture_lod');
  //ensureExt('EXT_sRGB');
  //ensureExt('EXT_texture_filter_anisotropic');
  ensureExt('OES_element_index_uint');
  //ensureExt('OES_standard_derivatives');
  ensureExt('OES_texture_float');
  ensureExt('OES_texture_float_linear');
  ensureExt('OES_texture_half_float');
  ensureExt('OES_texture_half_float_linear');
  ensureExt('OES_vertex_array_object');
  //ensureExt('WEBGL_color_buffer_float');
  //ensureExt('WEBGL_compressed_texture_atc');
  //ensureExt('WEBGL_compressed_texture_es3');
  //ensureExt('WEBGL_compressed_texture_etc1');
  //ensureExt('WEBGL_compressed_texture_pvrtc');
  //ensureExt('WEBGL_compressed_texture_s3tc');
  //ensureExt('WEBGL_depth_texture');
  //ensureExt('WEBGL_draw_buffers');
  ensureExt('WEBGL_lose_context');

  // Draft extensions, which should not be exposed by default.
  ensureNoExt('EXT_disjoint_timer_query');
  ensureNoExt('WEBGL_compressed_texture_es3');

  // Not implemented.
  ensureNoExt('EXT_color_buffer_float');
  ensureNoExt('OES_fbo_render_mipmap');
  ensureNoExt('WEBGL_compressed_texture_astc');
  ensureNoExt('WEBGL_security_sensitive_resources');
  ensureNoExt('WEBGL_shared_resources');

  // Privileged
  //ensureExt('WEBGL_debug_renderer_info');
  //ensureExt('WEBGL_debug_shaders');
} while (false);

    </script>
  </body>
</html>
